# openapi/v1/api.yaml
openapi: "3.0.2"
info:
  title: Elite Tapp API
  version: "1.0.0"
servers:
  - url: http://localhost:5000/api/v1
    description: Local
  - url: https://elite-tapp-api.herokuapp.com/api/v1
    description: Heroku Production
  - url: https://elitetapp-api.herokuapp.com/api/v1
    description: Syeta Production
...
paths:
...
  /devices/{id}:
    patch:
      summary: Update device
      description: Update device
      operationId: updateDevice
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      tags:
        - Devices
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: My Cool Tag
                  description: name of the device
                  nullable: true
                active:
                  type: boolean
                  example: true
                  description: Set device status
                  nullable: true
      responses:
        "200":
          $ref: "#/components/responses/OkResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnknownError"
        "500":
          $ref: "#/components/responses/UnknownError"
    delete:
      summary: Delete device
      description: Delete the device from system
      operationId: deleteDevice
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      tags:
        - Devices
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/OkResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnknownError"
        "500":
          $ref: "#/components/responses/UnknownError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    OkResponse:
      description: Ok
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MessageResponse"
...
    UnknownError:
      description: Unknown Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
    ValidationError:
      description: ValidationError Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
  schemas:
...
    MessageResponse:
      type: object
      description: message
      properties:
        message:
          type: string
          example: ok
...
    APIError:
      allOf:
        - $ref: "#/components/schemas/ErrorBase"
      properties:
        errors:
          type: array
          title: an array of errors
          items:
            type: object
    ErrorBase:
      type: object
      properties:
        name:
          type: string
          title: Name of the error
        message:
          type: string
          title: Descriptive error message
        status:
          type: integer
          title: HTTP Error code
        errorCode:
          type: integer
          title: API Specific error code
...
  parameters:
    IdPathParam:
      in: path
      name: id
      schema:
        type: string
        example: 8e9d94dd-6c04-48bb-882e-01f77f5224e9
      required: true
...